+++ cypress/e2e/data-export/display-list-of-settings-pages-permissions.cy.js
@@ -1,19 +1,22 @@
-import permissions from '../../support/dictionary/permissions';
+import Permissions from '../../support/dictionary/permissions';
 import TopMenu from '../../support/fragments/topMenu';
+import DevTeams from '../../support/dictionary/devTeams';
+import TestTypes from '../../support/dictionary/testTypes';
+import UsersSearchPane from '../../support/fragments/users/usersSearchPane';
 import Users from '../../support/fragments/users/users';
-import devTeams from '../../support/dictionary/devTeams';
-import testTypes from '../../support/dictionary/testTypes';
-import dataExportPermission from '../../support/fragments/data-export/data-export-permission';
+import UserEdit from '../../support/fragments/users/userEdit';
 
 let user;
 
-describe('Data-export', () => {
+describe('data-export', () => {
   before('Create test data', () => {
-    cy.createTempUser([permissions.uiUsersView.gui, permissions.uiUsersPermissions.gui]).then(
+    cy.createTempUser([Permissions.uiUsersView.gui, Permissions.uiUsersPermissions.gui]).then(
       (userProperties) => {
         user = userProperties;
-        cy.login(user.username, user.password);
-        cy.visit(TopMenu.dataExportPath);
+        cy.login(user.username, user.password, {
+          path: TopMenu.usersPath,
+          waiter: UsersSearchPane.waitLoading,
+        });
       },
     );
   });
@@ -24,24 +27,21 @@ describe('Data-export', () => {
 
   it(
     'C389473 Verify that "Settings (Data export): display list of settings pages" permission is searchable and renamed (firebird) (Taas)',
-    { tags: [devTeams.firebird, testTypes.extendedPath] },
+    { tags: [DevTeams.firebird, TestTypes.extendedPath] },
     () => {
-      dataExportPermission.openUsersMenu();
-      dataExportPermission.selectActiveStatus();
-      dataExportPermission.openUsersProfile();
-      dataExportPermission.verifyUserDetailsPane();
-      dataExportPermission.goToEditPage();
-      dataExportPermission.verifyUserPermissionsAccordion();
-      dataExportPermission.openSelectPermissions();
-      dataExportPermission.searchForPermission(
-        'Settings (Data export): display list of settings pages',
-      );
-      dataExportPermission.verifyPermissionsFiltered([
+      UsersSearchPane.searchByStatus('Active');
+      UsersSearchPane.searchByUsername(user.username);
+      Users.verifyUserDetailsPane();
+      Users.editButton();
+      UserEdit.verifyUserPermissionsAccordion();
+      UserEdit.openSelectPermissions();
+      UserEdit.searchForPermission('Settings (Data export): display list of settings pages');
+      UserEdit.verifyPermissionsFiltered([
         'Settings (Data export): display list of settings pages',
       ]);
-      dataExportPermission.resetAll();
-      dataExportPermission.searchForPermission('Data export');
-      dataExportPermission.verifyPermissionsFiltered([
+      UserEdit.resetAll();
+      UserEdit.searchForPermission('Data export');
+      UserEdit.verifyPermissionsFiltered([
         'Settings (Data export): display list of settings pages',
         'UI: Data export module is enabled',
       ]);
